+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdStrikeFalconAdversaryIntelligence/function_app.py
@@ -1,0 +1,13 @@
+import azure.functions as func
+from CrowdStrikeFalconThreatIntelConnector.main import run
+
+app = func.FunctionApp()
+
+
+@app.timer_trigger(
+    schedule="0 */10 * * * *",
+    arg_name="mytimer",
+    run_on_startup=False,
+)
+def CrowdStrikeFalconThreatIntelConnector(mytimer: func.TimerRequest) -> None:
+    run(mytimer)
