+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdStrikeFalconAdversaryIntelligence/requirements.txt
@@ -1,3 +1,3 @@
-azure-functions>=1.18.0
+azure-functions>=1.21.0
 requests>=2.32.4
 azure-storage-file-share>=12.22.0
